requirements.txt
# Flask Web Framework
Flask==2.3.3
Werkzeug==2.3.7

# Machine Learning
tensorflow==2.13.0
xgboost==2.0.0
scikit-learn==1.3.2

# Data Processing
pandas==2.1.4
numpy==1.24.3
scipy==1.11.4

# Financial Data
yfinance==0.2.40

# Visualization
matplotlib==3.8.2
plotly==5.17.0

# Utilities
python-dotenv==1.0.0
joblib==1.3.2

# Production Deployment
gunicorn==21.2.0

# System Requirements: Python 3.8+
